Completely rewrite and significantly simplify
This library used to be a wrapper around mkcollector. When attempting
to wrap it using FFI wrappers in MK, I realized it was simpler to just
use mkcollector for resubmission and to heavily refactor this library
to just allow for external tests, e.g. Psiphon.

It took me some more reasoning and some more hours of work to nail
down the proper structure for this library.

This commit effectively restructures the code as seems best to me and
additionally makes sure we have unit tests for the code.